First cut at a general container using runit to run the service.
In addition to running whatever is in the /etc/service directory it
will run anything specified on the command line (or with CMD)
with stdin and stdout from docker and terminating the runsvdir
process when the command exits. This replicates expected functionality
when using the container to run something that errors out or is
expected to terminate.